cambio de estilo el portfolio1
