+++ package.json
@@ -19,7 +19,8 @@
     "react-dom": "^18.2.0",
     "react-scripts": "5.0.1",
     "react-slick": "^0.30.2",
-    "slick-carousel": "^1.8.1"
+    "slick-carousel": "^1.8.1",
+    "swiper": "^11.0.6"
   },
   "devDependencies": {
     "@babel/plugin-proposal-private-property-in-object": "*",
