+++ src/App.js
@@ -1,20 +1,25 @@
 import "./css/navBar.css";
-import "./css/aboutMe.css";
-import "./css/header.css";
 import "./css/background.css";
 import "./css/contact.css";
 import "./css/icons.css";
 import "./css/proyects.css";
-import "./css/description.css";
-
+import "./css/description.css"
+import "./css/aboutme1.css";
 import "bootstrap/dist/css/bootstrap.min.css";
+
+
 import Header from "./components/header";
-import Aboutme from "./components/aboutme";
 import Proyects from "./components/proyects";
 import Info from "./components/icons";
 import Navi from "./components/NavBar";
-import Description from "./components/description";
+// import Slides from "./components/Slides";
+import Aboutme from "./components/aboutMe1";
+import Description from "./components/description1";
+
 import { Player } from "@lottiefiles/react-lottie-player";
+
+
+
 
 function App() {
   return (
@@ -22,16 +27,16 @@ function App() {
       <section className="back__img">
         <div>
           <Navi />
-          <Header />
           <main className="content" id="Content" data-scroll="area"></main>
           <div className="content_inner" data-scroll="target"></div>
+          <Description />
+           <Proyects /> 
+          <div className="back-aboutme">
           <Aboutme />
-        
-          <Proyects />
-          {/* <Description /> */}
+          
           <Info />
-        
         </div>
+         </div>
       </section>
     </>
   );
