+++ src/components/Slides.jsx
@@ -1,0 +1,38 @@
+// import React, { useRef, useState } from 'react';
+// // Import Swiper React components
+// import { Swiper, SwiperSlide } from 'swiper/react';
+
+// // Import Swiper styles
+// import 'swiper/css';
+// import 'swiper/css/pagination';
+// import 'swiper/css/navigation';
+
+// import './App.css';
+
+// // import required modules
+// import { Pagination, Navigation } from 'swiper/modules';
+
+// export default function Slides() {
+//   return (
+//     <>
+//       <Swiper
+//         pagination={{
+//           type: 'progressbar',
+//         }}
+//         navigation={true}
+//         modules={[Pagination, Navigation]}
+//         className="mySwiper"
+//       >
+//         <SwiperSlide>Slide 1</SwiperSlide>
+//         <SwiperSlide>Slide 2</SwiperSlide>
+//         <SwiperSlide>Slide 3</SwiperSlide>
+//         <SwiperSlide>Slide 4</SwiperSlide>
+//         <SwiperSlide>Slide 5</SwiperSlide>
+//         <SwiperSlide>Slide 6</SwiperSlide>
+//         <SwiperSlide>Slide 7</SwiperSlide>
+//         <SwiperSlide>Slide 8</SwiperSlide>
+//         <SwiperSlide>Slide 9</SwiperSlide>
+//       </Swiper>
+//     </>
+//   );
+// }
