+++ src/components/header.jsx
@@ -3,7 +3,7 @@ function Header (props) {
         <header className="siteHeader" id="SiteHeader">
       <div className="Header__t"> 
        <h1 className="siteHeader_title1">Hi!</h1> <h1 className="siteHeader_title"> I'm  Sofía Fornaroli.</h1>
-        <p className="siteHeader_label">I'm a Frontend Developer</p>  
+        {/* <p className="siteHeader_label">I'm a Frontend Developer</p>   */}
         </div>
 
         </header>
