--- src/css/aboutMe.css
@@ -1,62 +1,0 @@
-.home_aboutme {
-  width: 755px;
-  padding: 3rem 2rem;
-  display: flex;
-  justify-content: center;
-
-  color: #fff;
-  letter-spacing: 1px;
-  line-height: 2.5rem;
-  /* text-align: center; */
-  font-family: "Montserrat", sans-serif;
-  font-size: 22px;
-  font-weight: 400;
-
-}
-
-.lf-player-container{
-  margin: -25px;
-}
-
-.myproyects_title {
-
-  text-align: center;
-  align-items: center;
-  font-family: "Inter Tight", sans-serif;
-  display: flex;
-  justify-content: center;
-  font-weight: 800;
-  background: #80d7ff;
-  background: linear-gradient(to left, #80d7ff 45%, #ad56dd 66%);
-  background-clip: text;
-  -webkit-text-fill-color: transparent;  display: flex;
-  flex-direction: row;
-  flex-wrap: wrap;
-  justify-content: space-around;
-  list-style: none;
-
-  
-}
-
-.page_content {
-  width: 100%;
-  display: flex;
-  text-align: left;
-  justify-content: space-evenly;
-}
-
-
-@media screen and (max-width: 850px) {
-.arg__imagen{
-  height: 124px;
-}
-
-.page_content{
-  align-items: center;
-}
-
-.siteHeader_label{
-  width: 90%;
-}
-  
-}
