+++ src/css/background.css
@@ -1,10 +1,14 @@
 .back__img
+
 {
-    background: radial-gradient(circle at 30% 33%, rgba(0, 0, 0, 1) 0%, rgba(0, 0, 0, 1) 8%, rgba(0, 0, 0, 1) 15%, rgba(0, 0, 0, 1) 20%, rgba(0, 0, 0, 1) 37%, rgba(13, 13, 13, 0.97) 77%);
+    /* global 94%+ browsers support */
+    background: linear-gradient(187deg, rgba(8, 3, 43, 1) 0%, rgba(0, 0, 0, 1) 18%);
 }
-    /* {
-        background: radial-gradient(circle at 0% 41%, rgba(0, 0, 0, 1) 0%, rgba(0, 0, 0, 1) 8%, rgba(0, 0, 0, 1) 15%, rgba(0, 0, 0, 1) 26%, rgba(0, 0, 0, 1) 34%, rgba(0, 0, 0, 1) 43%, rgba(0, 0, 0, 1) 48%, rgba(0, 0, 0, 0.92) 100%);
 
 
+
+ .back-aboutme
+ {
+    /* global 94%+ browsers support */
+    background: linear-gradient(187deg, rgba(0, 0, 0, 1) 0%, rgba(0, 0, 0, 1) 60%, rgba(6, 2, 33, 1) 100%);
 }
- */
