+++ src/css/description.css
@@ -1,21 +1,83 @@
-.description_p {
-    width: 60%;
-  
-    color: #fff;
-    /* text-align: center; */
-    letter-spacing: 2px;
+.home_description {
+  width: 755px;
+  padding: 1rem 2rem;
+  display: flex;
+  flex-direction: column;
+
+  color: #fff;
+  letter-spacing: 1px;
+  line-height: 2.5rem;
+  font-family: "Montserrat", sans-serif;
+  font-size: 40px;
+  font-weight: 400;
+}
+
+.home_name {
+  font-family: "Inter Tight", sans-serif;
+  font-size: 80px;
+  font-weight: normal;
+  color: rgb(242, 242, 254);
+  margin-bottom: 2rem;
+}
+
+.player {
+  overflow: visible;
+  align-self: self-end;
+}
+
+.lf-player-container {
     display: flex;
-    justify-content: center;
-    font-family: "Montserrat", sans-serif;
-    font-size: 18px;
-    font-weight: 400;
-    padding: 2rem;
+}
+
+.myproyects_title {
+  text-align: center;
+  font-family: "Inter Tight", sans-serif;
+  display: flex;
+  list-style: none;
+
+  font-weight: 800;
+  background: #80d7ff;
+  background: linear-gradient(to left, #80d7ff 45%, #ad56dd 66%);
+  background-clip: text;
+  -webkit-text-fill-color: transparent;
+  
+  display: flex;
+  flex-direction: row;
+  flex-wrap: wrap;
+  justify-content: center;
+  align-items: center;
+
+}
+
+.page {
+  display: flex;
+  padding: 1rem 2rem;
+  margin-top: 100px;
+  width: 100%;
+  justify-content: center;
+}
+
+.animation {
+  display: flex;
+  justify-content: center;
+}
+
+.css-1c2fuzs {
+  animation-delay: 0ms;
+  display: flex;
+  justify-content: center;
+}
+
+@media screen and (max-width: 850px) {
+  .arg__imagen {
+    height: 124px;
   }
-  
-  
-  .description_div {
-    width: 100%;
-    display: flex;
-    /* justify-content: center; */
+
+  .animation {
+    align-items: center;
   }
-  
+
+  .siteHeader_label {
+    width: 90%;
+  }
+}
