+++ src/css/icons.css
@@ -2,7 +2,7 @@
     display: flex;
     flex-direction: row;
     justify-content: center;
-    padding: 70px 120px;
+    padding: 70px 120px 26px 120px;
 
 }
 
@@ -11,7 +11,7 @@
 }
 
 .instagram_icon, .github_icon, .email_icon, .linkedin_icon{
-    width: 65px;
-    margin: 0 2rem;
+    width: 37px;
+    margin: 0 15px;
 }
